Convert buildEnv.mjs -> buildEnv.js
